agent-service/requirements.txt
# CareTopicz Agent Service - Dependencies
# Python 3.11+

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Config
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Async HTTP
httpx>=0.28.0

# LangGraph + Claude
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0

# Testing
pytest>=8.0.0
